Feature/development/2.x

See merge request !2394

+++ kadena2.0/Kadena2.0.Dto/ViewOrder/MicroserviceResponses/History/ItemDto.cs
@@ -4,7 +4,7 @@
     {
         public string ArtWorkKey { get; set; }
         public string ClientId { get; set; }
-        public int DateTimeDeleted { get; set; }
+        public long DateTimeDeleted { get; set; }
         public int IsDeleted { get; set; }
         public string ItemId { get; set; }
         public string ItemName { get; set; }
